Convert home Form into a Component
